Refactor AssertVertex into Util packaged
