+++ keanu-project/src/main/java/io/improbable/keanu/vertices/utility/AssertVertex.java
@@ -1,6 +1,10 @@
-package io.improbable.keanu.vertices;
+package io.improbable.keanu.vertices.utility;
 
 import io.improbable.keanu.tensor.bool.BooleanTensor;
+import io.improbable.keanu.vertices.LoadVertexParam;
+import io.improbable.keanu.vertices.NonProbabilistic;
+import io.improbable.keanu.vertices.SaveVertexParam;
+import io.improbable.keanu.vertices.Vertex;
 import io.improbable.keanu.vertices.bool.BoolVertex;
 import io.improbable.keanu.vertices.dbl.KeanuRandom;
